// src/types/Mint.types.ts
// Generated by https://quicktype.io

export interface MintData {
  iowner: string;
  iamount: string;
  itokenid: string;
  iroyalties: Iroyalty[];
  itokenMetadata: ItokenMetadata;
}

export interface Iroyalty {
  partValue: string;
  partAccount: string;
}

export interface ItokenMetadata {
  ipfs: string;
  price: string;
  title: string;
  token: string;
  license: string;
  category: string;
  quantity: string;
  royalties: string;
  description: string;
}